Split gRPC usage to wrapper library
