--- src/grpc_exp.h
@@ -1,8 +1,0 @@
-//
-// Created by vincent on 3/12/23.
-//
-
-#ifndef PIDGIN_STEAM_GRPC_EXP_H
-#define PIDGIN_STEAM_GRPC_EXP_H
-
-#endif //PIDGIN_STEAM_GRPC_EXP_H
